### 2023-04-14 #5370 @editorialbot
[PRE REVIEW]: Ethical Smart Grid: a Gym environment for learning ethical behaviours
**Submitting author:** @rchaput (<a href="http://orcid.org/0000-0002-2233-7566">Rémy Chaput</a>)
**Repository:** https://github.com/ethicsai/ethical-smart-grid
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
