### 2023-04-17 ae17810
Added notebook to generate the OpenEI data profiles
Although the OpenEI data profiles are already in the repository
(`data/openei`), it was not clear how they were created from the
OpenEI full dataset.
This commit adds a notebook that demonstrates from scratch how to
generate these files (downloading the dataset files, extracting them,
parsing, and finally writing the `.npz` files).

It was important for:
- transparency;
- reproducibility;
- giving an example for other datasets.

Note that the existing data profiles are not re-created; but users can
use this notebook to re-generate and verify 
[…]
