### 2023-04-17 556400a
Fixed DOIs in the paper
As per the editorialbot (https://github.com/openjournals/joss-reviews/issues/5370#issuecomment-1508010439):

* Added missing DOIs for 5 references.
* Removed the 'https://doi.org/' prefix in 1 reference.
