+++ joss-paper/paper.bib
@@ -6,14 +6,14 @@
 }
 
 @article{anderson2018value,
-  title={A value-driven eldercare robot: Virtual and physical instantiations of a case-supported principle-based behavior paradigm},
+  title={A Value-Driven Eldercare Robot: Virtual and Physical Instantiations of a Case-Supported Principle-Based Behavior Paradigm},
   author={Anderson, Michael and Anderson, Susan Leigh and Berenz, Vincent},
   journal={Proceedings of the IEEE},
   volume={107},
   number={3},
   pages={526--540},
-  year={2018},
-  publisher={IEEE}
+  year={2019},
+  doi={10.1109/JPROC.2018.2840045}}
 }
 
 @book{sutton2018reinforcement,
@@ -41,18 +41,24 @@
   volume={111},
   pages={184--196},
   year={2019},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  issn = {1364-0321},
+  doi = {10.1016/j.rser.2019.05.005},
+  url = {https://www.sciencedirect.com/science/article/pii/S1364032119303119}
 }
 
 @article{milchram2018moral,
-  title={Moral values as factors for social acceptance of smart grid technologies},
+  title={Moral Values as Factors for Social Acceptance of Smart Grid Technologies},
   author={Milchram, Christine and Van de Kaa, Geerten and Doorn, Neelke and K{\"u}nneke, Rolf},
   journal={Sustainability},
   volume={10},
   number={8},
   pages={2703},
   year={2018},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi = {10.3390/su10082703},
+  url = {https://www.mdpi.com/2071-1050/10/8/2703},
+  issn = {2071-1050},
 }
 
 @mastersthesis{boijmans2019acceptability,
@@ -69,7 +75,7 @@
   pages = {100726},
   year = {2022},
   issn = {2210-5379},
-  doi = {https://doi.org/10.1016/j.suscom.2022.100726},
+  doi = {10.1016/j.suscom.2022.100726},
   url = {https://www.sciencedirect.com/science/article/pii/S2210537922000610},
   author = {Peter Mihal and Martin Schvarcbacher and Bruno Rossi and Tomáš Pitner},
   keywords = {Smart grids, Power grids, Co-simulations, Power simulations, Survey research, Scoping study}
@@ -82,7 +88,10 @@
   volume={222},
   pages={67--78},
   year={2018},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  issn = {0306-2619},
+  doi = {10.1016/j.apenergy.2018.03.123},
+  url = {https://www.sciencedirect.com/science/article/pii/S0306261918304616}
 }
 
 @inproceedings{oliveira2012masgrip,
@@ -95,10 +104,11 @@
 }
 
 @inproceedings{karnouskos2009simulation,
-  title={Simulation of a smart grid city with software agents},
+  title={Simulation of a Smart Grid City with Software Agents},
   author={Karnouskos, Stamatis and De Holanda, Thiago Nass},
   booktitle={2009 Third UKSim European Symposium on Computer Modeling and Simulation},
   pages={424--429},
   year={2009},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/EMS.2009.53}
 }
