### 2023-04-18 #5370 @kthyng
@rchaput On the editor side, we are trying to identify what direction to come from — the energy (smart grid) side vs the ethical side. Do you have any input or advice on this?
