### 2023-04-18 #5370 @rchaput
It is difficult to give a definitive answer, but I would say that the ethical side was essential in this work, in the sense that we built a smart grid simulator *for learning ethics*, not "for the sake of it". In addition, the need for such a simulator principally comes from the Machine Ethics field, rather than the Energy field (which already has several simulators).
