+++ docs/source/api.rst
@@ -2,7 +2,7 @@ API Reference
 =============
 
 .. image:: /images/architecture.drawio.png
-   :target: /images/architecture.drawio.png
+   :target: ./_images/architecture.drawio.png
    :alt: Simplified class diagram representing the links between the important classes. Attributes and methods are not shown.
 
 .. autosummary::
