### 2023-05-22 aedcad1
Add previous storage in AgentState
Accessing the previous storage is necessary for one of the arguments in the argumentation graphs (argumentation-based reward functions).
This commit thus opens the way to these new reward functions.
