### 2023-05-22 9a3b532
Added the argumentation-based reward functions
Added the following new classes that derive from `Reward`:
* Inclusiveness
* EnvironmentalSustainability
* Affordability
* SupplySecurity

Because the various moral values are implemented in different classes,
they can be used either individually or at the same time.
However, the `parse_situation` helper must be called for each moral value,
which implies a small (time) cost.
