+++ smartgrid/rewards/argumentation/supply_security.py
@@ -1,0 +1,155 @@
+"""
+Reward based on the *security of supply* moral value, promoting well-being.
+"""
+
+from smartgrid.rewards import Reward
+from smartgrid.rewards.argumentation.lib import (
+    JudgingAgent, judgment, AFDM, Argument
+)
+from smartgrid.rewards.argumentation.situation import parse_situation
+
+
+class SupplySecurity(Reward):
+    """
+    Reward based on the *security of supply* moral value, promoting well-being.
+    """
+
+    name: str
+
+    judge: JudgingAgent
+    """
+    The JudgingAgent that represents the implemented moral value.
+
+    It contains the argumentation graph (AFDM) that stores the arguments and
+    attacks between them, as well as the transform function, and performs
+    the judgment over a given situation.
+    """
+
+    decision: str = 'moral'
+    """
+    Hardcoded decision to use for the judgment.
+
+    The framework we use (AFDM) supports multiple decisions, in this current
+    version we only need to decide whether the action was *moral* or not.
+    Thus, we use an hardcoded identifier for the decision: it must be the same
+    when creating :py:class:`~.Argument`\\s and when performing the judgment!
+    """
+
+    LOW_GLOBAL = 130_000
+    HIGH_GLOBAL = 190_000
+
+    def __init__(self, j_transform=None):
+        super().__init__()
+        if j_transform is None:
+            j_transform = judgment.j_simple
+        afdm = self._create_afdm()
+        self.judge = JudgingAgent('supply_security', afdm, j_transform)
+
+    def calculate(self, world, agent):
+        situation = parse_situation(world, agent)
+        return self.judge.judge(situation, self.decision)
+
+    def _create_afdm(self):
+        af = AFDM()
+
+        af.add_argument(Argument(
+            "good_comfort_1", "comfort above 94%",
+            lambda s: s["self_comfort"] > 0.94,
+            support=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "good_comfort_2", "comfort above 96%",
+            lambda s: s["self_comfort"] > 0.96,
+            support=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "average_comfort", "comfort above 91%",
+            lambda s: s["self_comfort"] > 0.91,
+            support=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "bad_comfort_1", "comfort less than 91%",
+            lambda s: s["self_comfort"] < 0.91,
+            counter=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "bad_comfort_2", "comfort less than 85%",
+            lambda s: s["self_comfort"] < 0.85,
+            counter=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "bad_comfort_3", "comfort less than 75%",
+            lambda s: s["self_comfort"] < 0.75,
+            counter=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "bias_1", "bias",
+            lambda s: True,
+            support=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "bias_2", "bias",
+            lambda s: True,
+            counter=[self.decision]
+        ))
+
+        af.add_argument(Argument(
+            "high_global", "global available energy is high",
+            lambda s: s["global_energy"] > self.HIGH_GLOBAL
+        ))
+
+        af.add_argument(Argument(
+            "low_global", "global available energy is low",
+            lambda s: s["global_energy"] < self.LOW_GLOBAL
+        ))
+
+        af.add_argument(Argument(
+            "school", "agent is a school",
+            lambda s: s["agent_type"] == "School"
+        ))
+
[…]
