+++ smartgrid/rewards/argumentation/__init__.py
@@ -1,3 +1,8 @@
 """
 Argumentation-based reward functions.
 """
+
+from .affordability import Affordability
+from .env_sustain import EnvironmentalSustainability
+from .inclusiveness import Inclusiveness
+from .supply_security import SupplySecurity

[…]
