### 2023-05-22 8efa3b1
Added Makefile target to remove the generated modules
The default `make clean` does not remove `source/modules`; however, since this directory is generated by Sphinx autosummary (e.g., when using `make html`), it can be deleted safely.
It might even be necessary to delete it, for example when the source code does no longer correspond to the generated rst files. Renaming a class might cause the documentation to crash when generating, because it will try to re-use exising generated files, which do not match the new class!
