+++ smartgrid/rewards/__init__.py
@@ -17,3 +17,4 @@ Other functions can also be based on symbolic reasoning. (Work in progress)
 
 from .reward import Reward
 from .reward_collection import RewardCollection
+from .reward_constraints import TimeConstrainedReward, AgentConstrainedReward
