### 2023-05-23 c66296e
Add possibility to enable and disable reward functions
Reward functions were previously always "enabled": they produce rewards at
all time steps, for all learning agents.
Some use cases could benefit from enabling/disabling them: changing the reward
functions for all agents depending on the time steps (i.e., "evolving social
mores"), or using distinct reward functions for learning agents (as is possible
in the MARL literature), etc.

This commit:
* adds a method `is_activated` in the `Reward` class, that can be used to
  determine whether the reward should be produced;
* checks whether the rew
[…]
