### 2023-06-06 ab8a235
Refactor World to access agents by names (dict)
* Added an `agents_by_name` attribute (dictionary) to access agents based on
  their name.
* Changed the previous `agents` attribute to a view over the dict's values.
  This should not change anything, as we can still iterate over this attribute.
* Constructed the dict and the view from the iterable of Agents passed in the
  constructor. This parameter was previously typed as a `List`, but was changed
  to `Iterable`, as we do not require it to be strictly a list. This does not
  change anything when using the class, as `Iterable` is a parent of 
[…]
