### 2023-06-08 c96d84e
Refactored SmartGrid to use PettingZoo (multi-agent) instead of Gymnasium
The Gymnasium API is inherently single-agent; we had tweaked it (in a similar
manner to an older version of MultiParticleEnvironment) to support multiple
agents. However, the "true" multi-agent API is PettingZoo, which is developed
by the same foundation (Farama) as Gymnasium.

This commit refactors the `SmartGrid` environment to inherit from PettingZoo's
`ParallelEnv` instead of the traditional Gymnasium's `Env`. `ParallelEnv`
refers to multiple agents acting at the same time in the environment (contrary
to the `ACE` wa
[…]
