+++ smartgrid/environment.py
@@ -1,22 +1,28 @@
 """
 The SmartGrid environment is the main entrypoint.
 """
-import random
 import warnings
+from typing import Optional, Dict, Tuple, Any, Iterable, List
 
-import gymnasium
 import numpy as np
 from gymnasium import Space
+import pettingzoo
+# Import a few type hints to make the documentation clearer
+from pettingzoo.utils.env import AgentID, ObsDict, ActionDict
 
-from smartgrid.agents import Action
-from smartgrid.rewards import RewardCollection
+from smartgrid.agents import Action, Agent
+from smartgrid.rewards import RewardCollection, Reward
 from smartgrid.world import World
 from smartgrid.observation import ObservationManager
 
-from typing import List, Optional
+# A few additional type hints to help (similar to PettingZoo's)
+InfoType = Dict[str, Any]
+InfoDict = Dict[AgentID, InfoType]
+RewardsType = Dict[str, float]
+RewardsDict = Dict[AgentID, RewardsType]
 
 
-class SmartGrid(gymnasium.Env):
+class SmartGrid(pettingzoo.ParallelEnv):
     """
     The SmartGrid environment is the main entrypoint.
 
@@ -40,18 +46,14 @@ class SmartGrid(gymnasium.Env):
         'render.modes': ['text'],
     }
 
-    action_space: List[Space]
+    _np_random: np.random.Generator
     """
-    The list of action spaces for all Agents.
-    """
+    The pseudo-random number generator (PRNG), for reproducibility.
 
-    observation_space: List[Space]
-    """
-    The list of observation spaces for all Agents.
-    
-    Because the observation space is in practice split between *global* and
-    *local* observations, this might not exactly correspond, please see
-    the :py:meth:`~smartgrid.environment.SmartGrid._get_obs` for details.
+    It should usually not be accessed by the user, and must be passed down to
+    elements of the SmartGrid (e.g., :py:class:`~.World`) that need it.
+    The generator is set by the :py:meth:`~.reset` method, optionally with a
+    specific seed.
     """
 
     observation_manager: ObservationManager
@@ -101,15 +103,15 @@ class SmartGrid(gymnasium.Env):
 
     def __init__(self,
                  world: World,
-                 rewards,
-                 max_step=None,
+                 rewards: List[Reward],
+                 max_step: int = None,
                  obs_manager: ObservationManager = None):
         """
         Create the SmartGrid environment.
 
         This sets most attributes of the environment, including the
         :py:attr:`~smartgrid.environment.SmartGrid.action_space` and
-        :py:attr:`~smartgrid.environment.SmartGrid.observation_space`.
+        :py:attr:`~smartgrid.environment.SmartGrid.observation_space`..
 
         .. warning::
             Remember that the env is not usable until you call :py:meth:`.reset` !
@@ -145,14 +147,21 @@ class SmartGrid(gymnasium.Env):
         self.reward_calculator = RewardCollection(rewards)
 
         # Configure spaces
-        self.action_space = []
-        self.observation_space = []
-        obs_space = self.observation_manager.observation.get_observation_space()
+        self.observation_spaces = {}
+        self.action_spaces = {}
         for agent in self.world.agents:
-            self.action_space.append(agent.profile.action_space)
-            self.observation_space.append(obs_space)
+            self.observation_spaces[agent.name] = obs_manager.observation.space(
+                self.world,
+                agent
+            )
+            self.action_spaces[agent.name] = agent.profile.action_space
 
-    def step(self, action_n):
+    def step(
+        self,
+        actions: ActionDict
+    ) -> Tuple[
+        ObsDict, RewardsDict, Dict[AgentID, bool], Dict[AgentID, bool], InfoDict
+    ]:
         """
[…]
