### 2023-06-08 8f6e0b5
Refactored Local and Global Observations to use dataclasses instead of namedtuples
The classes for the Observations used `namedtuple`, which was great to ensure
they are immutable, and that users can access their fields by names, e.g.,
`obs.comfort` (instead of having to find their index in a given list).

However, namedtuples bring 2 major disadvantages:
* they cannot be inherited (and thus extended) easily;
* only the fields' names are specified, not their types.

Instead, `dataclasses` are a more recent and better tool to create such classes
that are immutable, which fields can be accessed 
[…]
