+++ smartgrid/observation/global_observation.py
@@ -2,36 +2,32 @@
 Global observations of the World, shared by all Agents in the smart grid.
 """
 
-from collections import namedtuple
+import dataclasses
+from typing import ClassVar, Optional, Dict, Tuple
 
 import numpy as np
+from gymnasium.spaces import Space, Box
 
 from smartgrid.util import hoover
 
-global_fields = [
-    'hour',
-    'available_energy',
-    'equity',
-    'energy_loss',
-    'autonomy',
-    'exclusion',
-    'well_being',
-    'over_consumption',
-]
 
-
-class GlobalObservation(namedtuple('GlobalObservation', global_fields)):
+@dataclasses.dataclass(frozen=True)
+class GlobalObservation:
     """
     Global observations of the World, shared by all Agents in the smart grid.
 
-    These observations are not directly linked to a particular agent, but
+    Observations cannot be modified once created, to limit potential bugs.
+    Global observations are not directly linked to a particular agent, but
     rather to the whole society of agents in the :py:class:`.World`, i.e.,
     in this smart grid. Thus, the measures are the same for all agents.
 
     To optimize computations, we thus create global observations only once
-    each step.
+    each step. This is done through the :py:attr:`.last_step_compute` and
+    :py:attr:`.computed` *class* attributes. We emphasize that they should
+    **not** be accessed through an instance, as they are not relevant as
+    observations, merely to *compute observations*.
 
-    A global observation is a vector containing the following measures:
+    A global observation contains the following measures:
 
     hour
         The current hour in the simulated world. It is computed as a ratio
@@ -89,7 +85,37 @@ class GlobalObservation(namedtuple('GlobalObservation', global_fields)):
         If the measure is less than 0, we set it to 0.
     """
 
-    last_step_compute = -1
+    # Instance attributes = observations measures
+
+    hour: float
+    """The current hour, represented in ``[0,1]``."""
+
+    available_energy: float
+    """The ratio of available energy in the Grid, compared to the maximum possible."""
+
+    equity: float
+    """The equity, a statistical measure of dispersion, between agents' comforts."""
+
+    energy_loss: float
+    """The ratio of energy not consumed over the total energy exchanges."""
+
+    autonomy: float
+    """
+    The ratio of exchanges that are not with the national grid, over the total exchanges.
+    """
+
+    exclusion: float
+    """The proportion of agents with a comfort less than half the median."""
+
+    well_being: float
+    """The median of agents' comforts."""
+
+    over_consumption: float
+    """The ratio of energy consumed that was not available, over the total exchanges."""
+
+    # Class attributes = used to memorize observations
+
+    last_step_compute: ClassVar[int] = -1
     """
     Last time step at which global observations were computed.
     
@@ -98,7 +124,7 @@ class GlobalObservation(namedtuple('GlobalObservation', global_fields)):
     time step.
     """
 
-    computed = None
+    computed: ClassVar[Optional['Self']] = None
     """
     Memoized global observations, computed at the time step indicated by
     :py:attr:`.last_step_compute`.
@@ -112,7 +138,7 @@ class GlobalObservation(namedtuple('GlobalObservation', global_fields)):
         return world.current_step == cls.last_step_compute
 
     @classmethod
-    def compute(cls, world: 'World'):
+    def compute(cls, world: 'World') -> 'Self':
         """
         Return the global observations computed from the World state.
 
@@ -173,15 +199,16 @@ class GlobalObservation(namedtuple('GlobalObservation', global_fields)):
         exclusion = len([c for c in comforts if c < threshold]) / len(comforts)
 
         cls.last_step_compute = world.current_step
-        cls.computed = cls(hour=hour,
-                           available_energy=available_energy,
-                           equity=equity,
[…]
