### 2023-06-26 ethicsai/ethical-smart-grid#1 @louiseadennis
Issue executing Quick Usage Instructions
I've installed ethical smart grid and the tests appear to run correctly.

However when I executed the Quick usage instructions I get a syntax error on

env.close()

I'm using Python 3.7.2.

Re: openjournals/joss-reviews#5410
