### 2023-07-11 ethicsai/ethical-smart-grid#1 @rchaput
Dear Pr. Dennis,

(Apologies for not answering earlier, it seems that I missed the notification)

When using IPython, the Quick usage works fine, but there is indeed a syntax error when using a "simple" Python REPL. A quick fix is to add a blank line between the end of the `while` loop and the `env.close()` line, like so:

```py
from smartgrid import make_basic_smartgrid
from algorithms.qsom import QSOM

env = make_basic_smartgrid(max_step=10)
model = QSOM(env)

done = False
obs = env.reset()
while not done:
    actions = model.forward(obs)
    obs, rewards, terminated, truncated, _ = env.step(actions)
    model.backward(obs, rewards)
    done = all(terminated) or all(truncated)

env.close()
```

(I also added the `max_step=10` argument to avoid running a 10k steps simulation and having the while loop hanging for ~ 1 hour, but this does not change anything with respect to the syntax error)

I will change the Readme instructions to avoid having a bug depending on the user's REPL.
