### 2023-07-18 ethicsai/ethical-smart-grid#1 @louiseadennis
Thanks!  It works now, though I do still get a warning!

`/opt/homebrew/lib/python3.11/site-packages/gymnasium/core.py:297: UserWarning: WARN: env.n_agent to get variables from other wrappers is deprecated and will be removed in v1.0, to get this variable you can do `env.unwrapped.n_agent` for environment variables or `env.get_attr('n_agent')` that will search the reminding wrappers.`

But it doesn't seem to stop anything working.
