### 2023-07-18 ethicsai/ethical-smart-grid#2 @louiseadennis
Community Guidelines
I'm not finding: "Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"

I think one can work it out, but maybe a statement at the end of the Readme?

Re: https://github.com/openjournals/joss-reviews/issues/5410#
