### 2023-07-18 ethicsai/ethical-smart-grid#3 @louiseadennis
Clarity of Documentation (Some Suggestions)
I found it initially difficult to understand how to use the code from the documentation.  In particular it would have helped if:

1. The quick usage example (or some other really simple example) could have printed some output.  Given the nature of the software the values representing Security of Supply, Affordability, Inclusiveness and Environmental Sustainability.  This would both give confidence that the example was actually doing something and give the user some immediate intuition about how algorithms could be compared.
2. The pre-programmed observations that represent these moral values should ideally be documented (or the documentation linked to) on the front page of the documentation where these values are mentioned.  This makes it clearer how the application actually supports these values.
3. A couple of pre-supplied scenarios with a couple of RL algorithms would be really useful, allowing a new user to run a comparison and observe the behaviour of the system before needing to adapt any code themselves.  They could also use these as a basis for creating their own experiments/scenarios without needing to build anything from scratch.
4. Ideally the pre-supplied elements would be better documented - perhaps in the API section which really only contains an overview diagram at present e.g., what pre-existing agent profiles exist? what generators? which reinforcement learning algorithms are supplied with the software; what are the pre-supplied reward functions; what are the pre-supplied observations.  There doesn't need to be a lot but a list for each of these with a sentence or two about each one.  In particular this would help people on first getting the system start quickly experimenting to see what they can do with it.

None of this is a deal-breaker, but my first reaction on having run the Quick Usage program was "what now?"

https://github.com/openjournals/joss-reviews/issues/5410#
