### 2023-07-24 8b8be18
Added contributing guidelines
Community guidelines were missing, both in the Readme and in a dedicated
file.
This commit adds the `CONTRIBUTING.md` file, which describes how to:

1. get support (ask questions);
2. report bugs;
3. propose changes.

The `Readme.md` also contains a short summary of these guidelines.

This commit fixes #2.
