### 2023-07-24 #5410 @louiseadennis
I'm happy to recommend Accept for this submission to JOSS.

I've left recommendations for some improvements to documentation as an issue, but believe that the submission is meets the JOSS criteria as it stands.
