### 2023-07-25 9ad6f8f
Update quick usage example to print rewards
The quick usage previously did not print *anything* (silent run). This was made
voluntarily, to avoid "cluttering" the standard output, but has the clear
disadvantage of not showing that "something is happening". It is a problem,
especially when using many time steps, which can be very long (e.g., several
minutes).

Instead, the quick usage was updated with 2 objectives in mind (in line with issue
#3): a) show that something is happening, and b) show the user intuitively how
algorithms could be compared.

Any basic message could be used to support ob
[…]
