### 2023-07-26 163e7be
Added notebook example to visualize rewards
Added an example notebook to show how to visualize "results", e.g.,
rewards received by learning agents when following a given algorithm.

This Jupyter Notebook relies on the MyST-NB extension, as it is a bit easier
to use than nbsphinx (which requires setting up pandoc and a kernel).

Note: we have hidden the first cell (using a tag `remove-input` in the cell
metadata), because it only pertains to our specific setup: the notebook is
built from the `docs/source/` folder, and cannot access the `smartgrid` and
`algorithms` modules. We had to change the
[…]
