+++ docs/source/usage.rst
@@ -71,7 +71,8 @@ The rewards received by the learning agents (``reward_n``) each step can be
 useful to analyze and visualize the quality of the learned behaviours.
 In this example, we simply print them, which has the additional advantage
 of showing that "something happens" at each time step; yet, most users will
-probably want to collect them and to display them in a plot.
+probably want to collect them and to display them in a plot. Please refer to
+:doc:`visualizing` to see an example of how to do so.
 
 In order to fully customize the environment setup, e.g., to control the
 number and profiles of agents, the available energy in the world, the reward
