### 2023-07-26 4437b62
Added pre-defined scenarii
Similar to the `make_basic_smartgrid`, added a few pre-defined scenarii that
have two objectives (in line with issue #3):

1. Allow users to quickly launch experiments, without having to code the
  environment setup themselves.
2. Provide inspiration for custom scenarii.

Note: these functions are heavy duplicates of the `make_basic_smartgrid`.
This is voluntary, as trying to share the different elements (creation of
agents, creation of world, selection of rewards, ...) would reduce the
readability of the functions, thus counter-intuitive to objective 2.
