### 2023-07-26 779c4bc
Allow building documentation on separate branch to test
We do not want to push to `master`, just in case the documentation does
not work (yet), so we push and build on another (dedicated) branch.
