+++ docs/source/api.rst
@@ -11,3 +11,140 @@ API Reference
 
    smartgrid
    algorithms
+
+
+To facilitate the exploration of the API, the pre-defined components of
+interest are listed below, along with a short description.
+
+- **Agents profiles and data conversion**
+
+  - :py:mod:`~smartgrid.agents.profile.comfort` functions : describe how the
+    agent's comfort should be computed, based on the consumption and need.
+
+    - :py:func:`~smartgrid.agents.profile.comfort.flexible_comfort_profile` :
+      A comfort function that is easy to satisfy: the curve quickly increases.
+
+    - :py:func:`~smartgrid.agents.profile.comfort.neutral_comfort_profile` :
+      A comfort function that increases "normally", the comfort follows more
+      or less the ratio ``consumption / need``.
+
+    - :py:func:`~smartgrid.agents.profile.comfort.strict_comfort_profile` :
+      A comfort function that is difficult to satisfy: the consumption must
+      be close to the need to see the comfort increase significantly.
+
+  - :py:mod:`Data Conversion <smartgrid.agents.data_conversion>` : describe
+    how to create an ``AgentProfile`` from raw data (dataset).
+
+    - :py:class:`~smartgrid.agents.data_conversion.DataOpenEIConversion` :
+      Creates profiles from the OpenEI dataset. Allows 3 kinds of buildings
+      (*Households*, *Offices*, *Schools*) and 2 sources of needs (*daily* -
+      aggregated and simplified, 24 data points ; *annual* - full dataset,
+      24*365 data points). Agent production is simulated (no available data).
+
+|
+
+- **World setup**
+
+  - :py:mod:`Energy Generators <smartgrid.util.available_energy>` : describe
+    how much energy should be available each step, based on the agents' total
+    need at this step, minimal and maximal need.
+
+    - :py:class:`~smartgrid.util.available_energy.RandomEnergyGenerator` :
+      The "basic" energy generator, simply returns a random value based on the
+      agents' current total need. The bounds are configurable (e.g., between
+      *80%* and *120%* of their current total need).
+
+    - :py:class:`~smartgrid.util.available_energy.ScarceEnergyGenerator` :
+      A random energy generator using *60%* and *80%* as bounds, i.e., there is
+      never enough (100%) energy to satisfy all agents' needs. This forces
+      conflict situations between agents.
+
+    - :py:class:`~smartgrid.util.available_energy.GenerousEnergyGenerator` :
+      A random energy generator using *100%* and *120%* as bounds, i.e., there
+      is always enough energy to satisfy all agents' needs. This represents an
+      easy scenario that can be combined with scarcity to see how agents
+      perform after a change (e.g., how agents trained for scarcity perform
+      during abundance, or conversely).
+
+    - :py:class:`~smartgrid.util.available_energy.RealisticEnergyGenerator` :
+      An energy generator that uses a dataset to determine how much energy
+      is generated at each step. To ensure that it scales with the number of
+      agents (and thus re-using datasets), this generator assumes that data
+      correspond to percentages of agents' maximum needs. This generator can be
+      used to introduce time-varying dynamics, e.g., less energy in winter than
+      in summer because of less solar power.
+
+|
+
+- **Reward functions**
+
+  - :py:mod:`Numerical reward functions <smartgrid.rewards.numeric>` : Reward
+    functions that are based on a "mathematical formula", as is traditionally
+    done in Reinforcement Learning.
+
+    - :py:mod:`Difference Rewards <smartgrid.rewards.numeric.differentiated>` :
+      Reward functions that are based on comparing the actual environment
+      with an hypothetical environment in which the agent did not act. This
+      gives us an idea of the agent's contribution.
+
[…]
