### 2023-07-28 fd930b1
Added link to rewards and observations supporting moral values
In the "Use-case" documentation page, we list the moral values that are
considered in this environment, but we did not specify how they were supported.
This commit (as per issue #3) adds a few lines that mention the reward functions
and observations that support these moral values, either by directly making
agents learn the expected behaviour (reward functions), or by providing a
measure about these moral values (observations). The mentioned reward functions
and observations are also linked to their own documentation.

Note: some m
[…]
