### 2023-07-28 ethicsai/ethical-smart-grid#3 @rchaput
Thank you for your feedback; I particularly appreciate this as, when writing documentation, it can be quite difficult to abstract all the details we already know about the software.
The 4 commits referenced here each target one of your proposed improvements; the issue is thus closed.
