### 2023-07-31 #5410 @seba-1511
Hello @ajstewartlang,

I'm also happy to accept this submission. I only have one comment for the authors:

The example in your GitHub README doesn't run out-of-the-box because the `algorithms` module is not included in `smartgrid`. I suggest either adding it or mentioning you need to add the directory to your python path.

Best,
