### 2023-08-04 2a0b88e
Added the algorithms package to the wheel release
Originally, this was left out voluntarily to avoid "cluttering"
the package environment of users. Thus, they could install `smartgrid`
from pip, and code their own algorithms.
However, users that install the package from pip instead of cloning
most likely want to start using it as soon as possible, and thus have
access to the `algorithms` package as well.
