+++ pyproject.toml
@@ -73,6 +73,6 @@ exclude = [
 
 # Also include the `data` folder, and move it to `smartgrid/data` when installed.
 [tool.hatch.build.targets.wheel]
-only-include = ["smartgrid", "data"]
+only-include = ["smartgrid", "data", "algorithms"]
 [tool.hatch.build.targets.wheel.sources]
 "data" = "smartgrid/data"
