+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "ethical-smart-grid"
-version = "1.0.0"
+version = "1.1.0"
 authors = [
     { name = "Clément Scheirlinck", email = "clement.scheirlinck@etu.univ-lyon1.fr" },
 ]
