### 2023-08-04 #5410 @rchaput
Sorry for the delay, I have just added the `algorithms` module to the built package (published on PyPi) as suggested.

Thank you for your review!

Please tell me if there is anything else.
