### 2023-08-10 #5410 @ajstewartlang
@rchaput Thanks for addressing the comments - I think we're almost there. In the paper, could you update the OpenAI  gym citation to please?

Brockman, G., Cheung, V., Pettersson, L., Schneider, J., Schulman, J., Tang, J., & Zaremba, W. (2016). Openai gym. ArXiv Preprint ArXiv:1606.01540.
