### 2023-08-11 49efbad
Fixed reference in paper
The reference for OpenAI Gym was missing the mention "arXiv preprint", as per
https://github.com/openjournals/joss-reviews/issues/5410
