+++ joss-paper/paper.bib
@@ -1,8 +1,8 @@
-@misc{gym,
-  Author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
-  Title = {OpenAI Gym},
-  Year = {2016},
-  Eprint = {arXiv:1606.01540}
+@article{gym,
+  title={OpenAI Gym},
+  author={Brockman, Greg and Cheung, Vicki and Pettersson, Ludwig and Schneider, Jonas and Schulman, John and Tang, Jie and Zaremba, Wojciech},
+  journal={arXiv preprint arXiv:1606.01540},
+  year={2016}
 }
 
 @article{anderson2018value,
