### 2023-08-11 #5410 @rchaput
Thank you for the review! I have updated the Gym citation.

For reference, I originally used the "official" citation from Gym's GitHub repository:
```bibtex
@misc{1606.01540,
  Author = {Greg Brockman and Vicki Cheung and Ludwig Pettersson and Jonas Schneider and John Schulman and Jie Tang and Wojciech Zaremba},
  Title = {OpenAI Gym},
  Year = {2016},
  Eprint = {arXiv:1606.01540},
}
```
(which is more or less the same as the ArXiv citation), but it seems that the JOSS style does not show that it is a pre-print.
Instead, the reference from Google Scholar does the job (but making it look like an article, and ArXiv a journal).
