### 2023-08-11 #5410 @rchaput
I have created a new [tagged release](https://github.com/ethicsai/ethical-smart-grid/releases/tag/v1.1.0-joss-paper) and archived it on Zenodo, here is the DOI:

[10.5281/zenodo.8239411](https://zenodo.org/record/8239411)

I have updated the metadata on Zenodo to make sure the title and authors correspond.
