### 2023-08-11 3aec2ca
Fixed missing label in paper
The link to "Open to extensions" did not work, because the label was not found, as per
https://github.com/openjournals/joss-reviews/issues/5410
Added an explicit label to the corresponding header.
