### 2023-08-25 #5410 @rchaput
Thank you for accepting our paper! I very much appreciated the JOSS editorial and reviewing process, it was great to receive frequent feedback. Many thanks again to the reviewers for their work.
